+++ test-suite/Math/NumberTheory/Powers/ModularTests.hs
@@ -53,11 +53,11 @@ powModProperty_Word (AnySign b) (NonNegative e) (Positive m) = powModWord b e m 
 
 -- | Specialized to trigger 'powModInteger'.
 powModProperty_Integer :: AnySign Integer -> NonNegative Integer -> Positive Integer -> Bool
-powModProperty_Integer (AnySign b) (NonNegative e) (Positive m) = powMod b e m == fromInteger (powMod' (fromIntegral b) (fromIntegral e) (fromIntegral m))
+powModProperty_Integer (AnySign b) (NonNegative e) (Positive m) = powMod b e m == fromInteger (powMod' b (fromIntegral e) m)
 
 -- | Specialized to trigger 'powModNatural'.
 powModProperty_Natural :: AnySign Natural -> NonNegative Natural -> Positive Natural -> Bool
-powModProperty_Natural (AnySign b) (NonNegative e) (Positive m) = powMod b e m == fromInteger (powMod' (fromIntegral b) (fromIntegral e) (fromIntegral m))
+powModProperty_Natural (AnySign b) (NonNegative e) (Positive m) = powMod b e m == fromInteger (powMod' (fromIntegral b) e (fromIntegral m))
 
 #if WORD_SIZE_IN_BITS == 64
 -- | Large modulo m such that m^2 overflows.
